restructuring of api calls
