+++ lib/providers/cart_provider.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:veganic_foods_app/screens/details_page/components/product_class.dart';
-import 'package:veganic_foods_app/utils/globals.dart';
 
 class Cart with ChangeNotifier {
   List<Product> _cart = [];
@@ -31,20 +30,46 @@ class Cart with ChangeNotifier {
   deletefromcart(index) {
     _cart.removeAt(index);
     notifyListeners();
-    final SnackBar snackbar = SnackBar(content: Text('item deleted'));
-    snackbarKey.currentState?.showSnackBar(snackbar);
+    Get.snackbar('', '',
+        messageText: Text('${_cart.elementAt(index).name} removed from the basket'),
+        titleText: Text(
+          'item  deleted',
+          style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),
+        ),
+        icon: Icon(Icons.remove),
+        padding: EdgeInsets.all(15),
+        backgroundColor: Colors.amberAccent.shade700);
   }
 
   addtoCart(Product product) {
-    if (_cart.contains(product)) {
-      Get.snackbar('yikes', 'item already exists bucko');
+    final itemExist =
+        _cart.where((element) => element.product_id == product.product_id);
+    if (itemExist.isNotEmpty) {
+      var addqt = _cart.firstWhere(
+          (element) => element.quantity == product.quantity,
+          orElse: null);
+      addqt.quantity = addqt.quantity + 1;
+      Get.snackbar('', '',
+          messageText: Text('${product.name} already exists'),
+          titleText: Text(
+            'oops',
+            style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),
+          ),
+          icon: Icon(Icons.playlist_add_check),
+          padding: EdgeInsets.all(15),
+          backgroundColor: Colors.amberAccent.shade700);
     } else {
       _cart.add(product);
       notifyListeners();
-
-      // final SnackBar snackbar = SnackBar(content: Text('item added'));
-      // snackbarKey.currentState?.showSnackBar(snackbar);
-
+      Get.snackbar('', '',
+          messageText: Text('${product.name} successfuly added to the basket'),
+          titleText: Text(
+            'item  added',
+            style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20),
+          ),
+          icon: Icon(Icons.playlist_add_check),
+          padding: EdgeInsets.all(15),
+          backgroundColor: Colors.amberAccent.shade700);
     }
   }
