+++ lib/screens/payment_page/payment.dart
@@ -3,12 +3,13 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 import 'package:provider/provider.dart';
 import 'package:veganic_foods_app/constants.dart';
-import 'package:veganic_foods_app/screens/payment_page/components/transaction_function.dart';
-import 'package:veganic_foods_app/utils/globals.dart';
 import 'package:veganic_foods_app/widgets/custom_button.dart';
 import '../../providers/cart_provider.dart';
 import '../../utils/routes.dart';
 import '../../widgets/bottom_nav_bar.dart';
+import '../../widgets/divider.dart';
+import '../../widgets/radiotile_css.dart';
+import '../../widgets/transaction_alert_dialog.dart';
 import 'components/background_eclipses.dart';
 
 class PaymentPage extends StatefulWidget {
@@ -62,7 +63,6 @@ class _PaymentListState extends State<PaymentList> {
                   Column(mainAxisAlignment: MainAxisAlignment.end, children: [
                 // ignore: prefer_const_constructors
                 backgroundbubbles(
-                  // height: size.height*0.1719.h,
                   height: height * 0.1,
                   name: '',
                 ),
@@ -87,7 +87,6 @@ class _PaymentListState extends State<PaymentList> {
                           borderRadius: BorderRadius.only(
                             topLeft: Radius.circular(40),
                             topRight: Radius.circular(40),
-                            // bottomLeft: Radius.elliptical(200,50 ),
                           )),
                       child: Column(children: [
                         SizedBox(
@@ -204,7 +203,6 @@ class _PaymentListState extends State<PaymentList> {
                               ),
                               if (context.read<Cart>().cart.isNotEmpty)
                                 Text(
-                                  // '${context.watch<Cart>().total}',
                                   '${Provider.of<Cart>(context, listen: false).total.toStringAsFixed(2)}',
                                   style: TextStyle(
                                       fontSize: 30,
@@ -295,197 +293,5 @@ class _PaymentListState extends State<PaymentList> {
   }
 }
 
-class TransactionAlertDalog extends StatelessWidget {
-  final String text;
-  final String validator;
-  final String hint;
-  const TransactionAlertDalog({
-    Key? key,
-    required GlobalKey<FormState> formkey,
-    required this.textcontroller,
-    required this.text,
-    required this.validator,
-    required this.hint,
-  })  : _formkey = formkey,
-        super(key: key);
 
-  final GlobalKey<FormState> _formkey;
-  final TextEditingController textcontroller;
 
-  @override
-  Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
-    //total height and width of screen
-
-    return AlertDialog(
-        shape: ShapeBorder.lerp(
-          RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(20),
-          ),
-          RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(20),
-          ),
-          0.5,
-        ),
-        backgroundColor: Colors.white,
-        title: Container(
-          alignment: Alignment.centerRight,
-          child: IconButton(
-              onPressed: () {
-                Navigator.pop(context);
-              },
-              icon: Icon(
-                Icons.close,
-                size: 30,
-                color: Colors.black,
-              )),
-        ),
-        content: Container(
-          height: size.height * 0.4,
-          padding: EdgeInsets.all(20),
-          alignment: Alignment.center,
-          child: Form(
-            key: _formkey,
-            child: Column(children: [
-              Text(text,
-                  style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold)),
-              SizedBox(
-                height: size.height * 0.02,
-              ),
-              TextFormField(
-                validator: (value) {
-                  if (value!.isEmpty || value.characters.length < 10) {
-                    return '$validator';
-                  } else {
-                    return null;
-                  }
-                },
-                style: TextStyle(color: Colors.black),
-                controller: textcontroller,
-                decoration: InputDecoration(
-                  hintText: '260 ',
-                  hintStyle: TextStyle(color: Colors.grey.shade700),
-                  labelText: hint,
-                  labelStyle: TextStyle(color: Colors.grey.shade700),
-                  focusedBorder: OutlineInputBorder(
-                    borderSide: BorderSide(color: bGcolor),
-                    borderRadius: BorderRadius.circular(10),
-                  ),
-                  border: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(10),
-                      borderSide: BorderSide(color: bGcolor)),
-                ),
-              ),
-              SizedBox(
-                height: size.height * 0.02,
-              ),
-              Text(
-                'your cart total is ${context.read<Cart>().total}',
-                style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-              ),
-              AppButton(
-                text: 'pay',
-                fontSize: 25,
-                textColor: Colors.white,
-                bgColor: Colors.black,
-                fontWeight: FontWeight.bold,
-                borderRadius: 30,
-                height: size.height * 0.02,
-                onTap: () {
-                  if (_formkey.currentState!.validate()) {
-                    gateway(
-                        textcontroller.text, context.read<Cart>().carttotal());
-                    Navigator.of(context).pop();
-                    final SnackBar snackBar = SnackBar(
-                        duration: Duration(seconds: 3),
-                        padding: EdgeInsets.only(top: 250, bottom: 300),
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(20),
-                        ),
-                        backgroundColor: Colors.transparent,
-                        content: AlertDialog(
-                          title: Text(''),
-                          content: Container(
-                            decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(70),
-                            ),
-                            alignment: Alignment.center,
-                            child: Column(
-            
-                              children: [
-                                if (context.read<Cart>().carttotal() != 0.0)
-                                  Icon(
-                                    Icons.check,
-                                    color: Colors.green,
-                                    size: 50,
-                                  ),
-                                Text(
-                                  'Transaction in progress',
-                                  style: TextStyle(
-                                      fontSize: 15,
-                                      fontWeight: FontWeight.bold),
-                                ),
-                              ],
-                            ),
-                          ),
-                        ));
-                    snackbarKey.currentState?.showSnackBar(snackBar);
-                  }
-                },
-              )
-            ]),
-          ),
-        ));
-  }
-}
-
-class divider extends StatelessWidget {
-  const divider({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Divider(
-      thickness: 1,
-      indent: 50,
-      endIndent: 50,
-    );
-  }
-}
-
-class RadiotileCSS extends StatelessWidget {
-  final String text;
-  final String imagestring;
-  const RadiotileCSS({
-    Key? key,
-    required this.text,
-    required this.imagestring,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
-    return Row(children: <Widget>[
-      InkWell(
-        child: Row(
-          // ignore: prefer_const_literals_to_create_immutables
-          children: [
-            Image.asset(
-              imagestring,
-              height: 25.h,
-            ),
-            SizedBox(
-              width: size.height * 0.02,
-            ),
-            Text(
-              text,
-              // ignore: prefer_const_constructors
-              style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-            )
-          ],
-        ),
-      ),
-    ]);
-  }
-}
