+++ lib/screens/scanning_page/components/https_service.dart
@@ -1,13 +1,10 @@
-import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
-import 'package:get/get.dart';
-import 'package:http/http.dart' as http;
 import 'package:veganic_foods_app/constants.dart';
+import 'package:veganic_foods_app/providers/Api_provider.dart';
 import 'package:veganic_foods_app/screens/details_page/details.dart';
-
+import '../../../widgets/network_error_page.dart';
 import '../../details_page/components/product_class.dart';
-import '../scan.dart';
 
 class Httpp extends StatefulWidget {
   final String? id;
@@ -18,12 +15,11 @@ class Httpp extends StatefulWidget {
 }
 
 class _HttppState extends State<Httpp> {
-  late Future<Product> _future ;
+  late Future<Product> _future;
   @override
   void initState() {
-    // TODO: implement initState
     super.initState();
-    _future = _getdata(widget.id);
+    _future = ApiProvider().getProduct(widget.id);
   }
 
   @override
@@ -84,61 +80,9 @@ class _HttppState extends State<Httpp> {
   }
 }
 
-const String postUrl = "http://192.168.235.5:8007/api/product";
-Future<Product> _getdata(String? id) async {
-  String url = postUrl + '/$id';
-  var res = await http.get(Uri.parse(url));
-  if (res.statusCode == 200) {
-//==================================================================
-    Map<String, dynamic> productMap = jsonDecode(res.body);
-    var products = Product.fromJson(productMap);
-    print(products.toJson());
-    return products;
-//====================================================================
-  } else
-    return Circle();
-  // return NetworkErrorpage();
-}
 
 dynamic Circle() {
   return CircularProgressIndicator();
 }
 
-dynamic NetworkErrorpage() {
-  return Scaffold(
-    body: Stack(
-      fit: StackFit.expand,
-      children: [
-        Image.asset(
-          "assets/images/1_No Connection.png",
-          fit: BoxFit.cover,
-        ),
-        Container(
-          padding:
-              const EdgeInsets.only(top: 100, left: 20, right: 20, bottom: 90),
-          alignment: Alignment.bottomCenter,
-          child: ElevatedButton(
-            style: ElevatedButton.styleFrom(
-              fixedSize: Size(300, 59),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(20),
-              ),
-              primary: Color(0xFF6B92F2),
-              onPrimary: Colors.white,
-            ),
-            onPressed: () {
-              Get.to(ScanningPage());
-            },
-            child: Text('scan again',
-                maxLines: 3,
-                style: TextStyle(
-                  fontSize: 30,
-                  fontWeight: FontWeight.bold,
-                  color: Colors.white,
-                )),
-          ),
-        )
-      ],
-    ),
-  );
-}
+
